[connect-ip] create a multiqueue device wrapper and performance test for connect-ip (#48)
Wow that sucks,

```
2025/05/23 13:01:25 INFO Sent packets sent_bytes=1887075840
2025/05/23 13:01:25 writing to stream failed: use of closed network connection
2025/05/23 13:01:25 handling stream failed: H3 error (0x0) (local): done
2025/05/23 13:01:25 handling stream failed: Application error 0x0 (remote): done
2025/05/23 13:01:25 writing to stream failed: use of closed network connection
PASS
```

Only like 1.5 Gbits on my little Dell XPS laptop. Will be a lot higher
on a modern ARM CPU but that's rather unimpressive. Looks like the
connect-ip library allocates on each packet or something. Also the lack
of exposing context, and the close behavior is pretty rubbish.

---------

Co-authored-by: Dmitry Ilyevsky <ilyevsky@gmail.com>